requirements.txt
langchain>=0.1.13
langchain-community>=0.0.30
sentence-transformers>=2.5
faiss-cpu>=1.7
streamlit>=1.34
python-dotenv>=1.0
pypdf>=4.0
groq>=0.5.0
langchain-groq>=0.1.2
